+++ src/main/java/com/iexec/sms/tee/TeeController.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020-2023 IEXEC BLOCKCHAIN TECH
+ * Copyright 2020-2024 IEXEC BLOCKCHAIN TECH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,13 +19,13 @@ package com.iexec.sms.tee;
 
 import com.iexec.common.web.ApiResponseBody;
 import com.iexec.commons.poco.chain.WorkerpoolAuthorization;
+import com.iexec.commons.poco.security.Signature;
 import com.iexec.commons.poco.tee.TeeFramework;
 import com.iexec.sms.api.TeeSessionGenerationError;
 import com.iexec.sms.api.TeeSessionGenerationResponse;
 import com.iexec.sms.api.config.TeeServicesProperties;
 import com.iexec.sms.authorization.AuthorizationError;
 import com.iexec.sms.authorization.AuthorizationService;
-import com.iexec.sms.tee.challenge.TeeChallenge;
 import com.iexec.sms.tee.challenge.TeeChallengeService;
 import com.iexec.sms.tee.session.TeeSessionService;
 import com.iexec.sms.tee.session.generic.TeeSessionGenerationException;
@@ -73,6 +73,7 @@ public class TeeController {
 
     /**
      * Return which TEE framework this SMS is configured to use.
+     *
      * @return TEE framework this SMS is configured to use.
      */
     @GetMapping("/framework")
@@ -100,15 +101,29 @@ public class TeeController {
     }
 
     /**
-     * Called by the core, not the worker
+     * Generates an enclave challenge for a PoCo task.
+     * <p>
+     * This method is called by the scheduler.
+     *
+     * @param authorization Authorization to check the query legitimacy
+     * @param chainTaskId   ID of the task the challenge will be produced for
+     * @return The Ethereum address enclave challenge for the provided
      */
     @PostMapping("/challenges/{chainTaskId}")
-    public ResponseEntity<String> generateTeeChallenge(@PathVariable String chainTaskId) {
-        Optional<TeeChallenge> executionChallenge =
-                teeChallengeService.getOrCreate(chainTaskId, false);
-        return executionChallenge
-                .map(teeChallenge -> ResponseEntity
-                        .ok(teeChallenge.getCredentials().getAddress()))
+    public ResponseEntity<String> generateTeeChallenge(@RequestHeader String authorization, @PathVariable String chainTaskId) {
+        log.debug("generateTeeChallenge [authorization:{}, chainTaskId:{}]", authorization, chainTaskId);
+        final Optional<AuthorizationError> authorizationError = authorizationService.isAuthorizedOnExecutionWithDetailedIssue(
+                WorkerpoolAuthorization.builder()
+                        .chainTaskId(chainTaskId)
+                        .enclaveChallenge("")
+                        .workerWallet("")
+                        .signature(new Signature(authorization))
+                        .build());
+        if (authorizationError.isPresent()) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+        return teeChallengeService.getOrCreate(chainTaskId, false)
+                .map(teeChallenge -> ResponseEntity.ok(teeChallenge.getCredentials().getAddress()))
                 .orElseGet(() -> ResponseEntity.notFound().build());
     }
 
@@ -118,10 +133,12 @@ public class TeeController {
      * to the enclave so the latter can talk to the CAS and get
      * the needed secrets.
      *
-     * @return
-     *      200 OK with the session id if success,
-     *      404 NOT_FOUND if the task is not found,
-     *      500 INTERNAL_SERVER_ERROR otherwise.
+     * @return result
+     * <ul>
+     * <li>200 OK with the session id if success.
+     * <li>404 NOT_FOUND if the task is not found.
+     * <li>500 INTERNAL_SERVER_ERROR otherwise.
+     * </ul>
      */
     @PostMapping("/sessions")
     public ResponseEntity<ApiResponseBody<TeeSessionGenerationResponse, TeeSessionGenerationError>> generateTeeSession(
@@ -140,7 +157,7 @@ public class TeeController {
                     .body(body);
         }
         final Optional<AuthorizationError> authorizationError =
-                authorizationService.isAuthorizedOnExecutionWithDetailedIssue(workerpoolAuthorization, true);
+                authorizationService.isAuthorizedOnExecutionWithDetailedIssue(workerpoolAuthorization);
         if (authorizationError.isPresent()) {
             final TeeSessionGenerationError teeSessionGenerationError =
                     authorizationToGenerationError.get(authorizationError.get());
@@ -170,7 +187,7 @@ public class TeeController {
             return ResponseEntity.ok(ApiResponseBody.<TeeSessionGenerationResponse, TeeSessionGenerationError>builder()
                     .data(teeSessionGenerationResponse)
                     .build());
-        } catch(TeeSessionGenerationException e) {
+        } catch (TeeSessionGenerationException e) {
             log.error("Failed to generate secure session [taskId:{}, workerAddress:{}]",
                     taskId, workerAddress, e);
             final ApiResponseBody<TeeSessionGenerationResponse, TeeSessionGenerationError> body =